Refactor: 게시판 댓글목록 컴포넌트 분리
